[Refactor] friend 리졸버 분리
합의한 리팩토링 방식에 맞춰 쿼리 뮤테이션, 서브스크라이브로 분리하였다.
