+++ server/src/api/friend/friendAlarm.subscription.ts
@@ -1,0 +1,15 @@
+import { withFilter } from 'graphql-subscriptions';
+
+const REQUEST_ALARM_ADDED = 'REQUEST_ALARM_ADDED';
+
+export const Subscription = {
+  requestAlarmAdded: {
+    subscribe: withFilter(
+      (_, __, { pubsub }) => {
+        return pubsub.asyncIterator(REQUEST_ALARM_ADDED);
+      },
+      async (payload, _, { email }) =>
+        payload.requestAlarmAdded.targetEmail === email
+    )
+  }
+};
